chore: 🤖 re-organized tree tests
